Update to reflect changes to `rhelder` and `rhelderhandout`
And in the process greatly simplify. First, assume that either 'rhelder'
or 'rhelderhandout' is going to be used. Do not try to reinvent the
wheel and make a template that can accommodate files that don't want to
call 'rhelder' or 'rhelderhandout'. Anyone who wants that can use the
default Pandoc template, which offers everything you need. The purpose
of the custom template should be to harness the power of
'rhelderhandout' or 'rhelder' to achieve a more concise preamble.

To that end, only pass biblatex styles and options to 'rhelderhandout'
or 'rhelder'. All other options are implemented via 'classoption' and
'rhelderoptions'. The one exception is the 'geometry' variable. If the
'geometry' variable is non-empty, call '\geometry' with the options
specified in the 'geometry' variable. 'geometry' is already loaded by
'rhelder', so this just makes it possible for tweaks to the geometry
settings (e.g., most often, to 'vscale') to be encoded in the Markdown
YAML header (as opposed to adjusted ad-hoc in the resultant 'tex' file).

Also:
*   Use 'rhhandout' as a variable instead of 'handout', which Pandoc
    already uses in its default template to produce a Beamer handout.
*   Add an 'include-after' block, which for some reason was ommitted in
    past versions.