Remove `display: block` property from `table`
It's not clear what work it was supposed to be doing, and it makes it
counterintuitive (or impossible?) to do simple tasks like centering the
table. This was part of the initial commit of 'styles.html' to Pandoc,
so no reason was given for its inclusion (also note that 'tex4ht' does
not style 'table' with 'display: block').